requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
aiofiles==23.2.1
tqdm==4.66.1
packaging==23.2
Pillow==10.1.0


